documentation-helper/pyproject.toml
[tool.poetry]
name = "documentation-helper"
version = "0.1.0"
description = ""
authors = ["jyoti <paljyoti0129@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "^4.12.3"
black = "^24.10.0"
tiktoken = "^0.8.0"
openai = "^1.58.1"
unstructured = "^0.16.11"
nltk = "^3.9.1"
fastapi = "^0.115.6"
jinja2 = "^3.1.4"
uvicorn = "^0.34.0"
streamlit = "^1.41.1"
streamlit-chat = "^0.1.1"
tqdm = "^4.67.1"
isort = "^5.13.2"
langchainhub = "^0.1.21"
pinecone-client = "^5.0.1"
pillow = "^11.0.0"
langchain-community = "^0.3.12"
langsmith = "^0.2.4"
firecrawl-py = "^1.6.4"
langchain = "^0.3.13"
langchain-cli = "^0.0.35"
langchain-pinecone = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
